fix(structure): fixed existing structure issues
