+++ src/auth/auth.service.ts
@@ -17,8 +17,8 @@ import {
 import { AllConfigType } from '@src/config/config.type';
 import { MailService } from '@src/mail/mail.service';
 import { RoleEnum } from '@src/roles/roles.enum';
-import { Session } from '@src/session/domain/session';
-import { SessionService } from '@src/session/session.service';
+import { Session } from '@src/sessions/domain/session';
+import { SessionService } from '@src/sessions/session.service';
 import { SocialInterface } from '@src/social/interfaces/social.interface';
 import { StatusEnum } from '@src/statuses/statuses.enum';
 import { User } from '@src/users/domain/user';
