+++ src/auth/strategies/types/jwt-payload.type.ts
@@ -1,4 +1,4 @@
-import { Session } from '@src/session/domain/session';
+import { Session } from '@src/sessions/domain/session';
 import { User } from '@src/users/domain/user';
 
 export type JwtPayloadType = Pick<User, 'id' | 'role'> & {
