+++ src/biometric-challenges/dtos/get-challenge-params.dto.ts
