+++ src/sessions/domain/session.ts
