+++ src/users/infrastructure/persistence/relational/relational-persistence.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { UserDeviceEntity } from '@src/users/infrastructure/persistence/relational/entities/user-device.entity';
 import { UserAbstractRepository } from '@src/users/infrastructure/persistence/user.abstract.repository';
 
 import { UserEntity } from './entities/user.entity';
 import { UsersRelationalRepository } from './repositories/user.repository';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UserEntity])],
+  imports: [TypeOrmModule.forFeature([UserEntity, UserDeviceEntity])],
   providers: [
     {
       provide: UserAbstractRepository,
